% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chek_reptiledb_status.R
\name{check_internet_connection}
\alias{check_internet_connection}
\title{Check Internet Connection}
\usage{
check_internet_connection()
}
\value{
Logical. TRUE if internet is available
}
\description{
Check Internet Connection
}
\keyword{internal}
