% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chek_reptiledb_status.R
\name{print.reptile_update_check}
\alias{print.reptile_update_check}
\title{Print method for reptile_update_check objects}
\usage{
\method{print}{reptile_update_check}(x, ...)
}
\arguments{
\item{x}{An object of class "reptile_update_check"}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for reptile_update_check objects
}
