% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chek_reptiledb_status.R
\name{check_data_update}
\alias{check_data_update}
\title{Check if reptile database data needs updating based on date comparison}
\usage{
check_data_update(silent = FALSE, check_connection = TRUE)
}
\arguments{
\item{silent}{Logical. If TRUE, suppresses messages and only returns results.
Default is FALSE.}

\item{check_connection}{Logical. If TRUE, checks internet connection before
attempting to access online data. Default is TRUE.}
}
\description{
This function checks if the local reptile database data is up-to-date by
comparing the date extracted from the local dataset name with the date
from the latest available file on The Reptile Database website.
}
\examples{
\dontrun{
# Silent check (no messages)
update_status <- check_data_update(silent = TRUE)
}

}
