% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chek_reptiledb_status.R
\name{extract_date_from_name}
\alias{extract_date_from_name}
\title{Extract date from dataset name or filename}
\usage{
extract_date_from_name(name, type = "local")
}
\arguments{
\item{name}{Dataset name or filename}

\item{type}{Type of name ("local" or "remote")}
}
\value{
Date object or NULL if extraction fails
}
\description{
Extract date from dataset name or filename
}
